requirements.txt
beautifulsoup4
requests
fpdf
pillow
python-dotenv
google-generativeai